Feed CrashesByYearCumulative with data from CrashesByYear, comment out old code
